fix some problems

+++ DDPackage/dd/Definitions.hpp
@@ -53,7 +53,7 @@ static constexpr fp PI_4 = static_cast<fp>(
     0.785398163397448309615660845819875721049292349843776455243L);
 
 
-static constexpr long int root_of_unit = 8;
+static constexpr int root_of_unit = 8;
 
 static constexpr double unit_rotate_angle = 2 * PI / root_of_unit;
